% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEMODMetrics.R
\name{calculate.bounds.two_sigma}
\alias{calculate.bounds.two_sigma}
\title{Calculate the 95 percent confidence interval of an estimate given 2 sigma. This is used to
calculate confidence intervals from EMOD calibration data that comes from a Gaussian distribution}
\usage{
calculate.bounds.two_sigma(mean_vector, two_sigma_vector)
}
\arguments{
\item{mean_vector}{A vector of the estimated value of a measurement}

\item{two_sigma_vector}{A vector of the two standard deviations from the mean of a measurement}
}
\value{
A data.frame with columns lb and ub corresponding to upper and lower bounds (95% CI)
}
\description{
Calculate the 95 percent confidence interval of an estimate given 2 sigma. This is used to
calculate confidence intervals from EMOD calibration data that comes from a Gaussian distribution
}
