% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEMODMetrics.R
\name{calculate.incidence}
\alias{calculate.incidence}
\title{Calculate incidence of an EMOD simulation}
\usage{
calculate.incidence(data)
}
\arguments{
\item{data}{A tibble returned from read.simulation.results(..., stratify_columns = c("Year", "Gender"), aggregate_columns = c("Population","Newly.Infected", "Infected")).
More aggregate columns can be used, but more stratify columns will be ignored}
}
\value{
A tibble with columns incidence and Year
}
\description{
Calculate incidence of an EMOD simulation
}
