% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEMODMetrics.R
\name{calculate.prevalence}
\alias{calculate.prevalence}
\title{Calculate prevalence in an EMOD simulation}
\usage{
calculate.prevalence(
  data,
  stratify_columns = c("Year", "Gender", "sim.id"),
  numerator = "Infected",
  denominator = "Population"
)
}
\arguments{
\item{data}{A tibble returned from read.simulation.results()}

\item{stratify_columns}{List of strings, names of stratifying columns included in the `group_by` function call before aggregation. The list should include sim.id (or sim.ix) as an argument to calculate the prevalence for each individual simulation run.}

\item{numerator}{String, name of prevalence numerator}

\item{denominator}{String, name of prevalence denominator}
}
\description{
Defaults to calculating HIV prevalence, but also may be used to calculate fraction of PLHIV who know their status, ART coverage, or other measures involving a numerator and denominator.
Prevalence = (numerator)/(denominator)
}
