% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEMODMetrics.R
\name{calculate.tests.performed}
\alias{calculate.tests.performed}
\title{Calculate number of tests performed in an EMOD simulation}
\usage{
calculate.tests.performed(data, gender.breakdown = TRUE, debug = FALSE)
}
\arguments{
\item{data}{A tibble returned from read.simulation.results(..., stratify_columns = c("Year", "Gender"), aggregate_columns = c("Population","Newly.Infected", "Infected")).
More aggregate columns can be used, but more stratify columns will be ignored}

\item{gender.breakdown}{- Boolean which controls whether to disaggregate the data by gender. Defaults to TRUE.}

\item{debug}{- Boolean which controls whether function returns intermediate results. Defaults to TRUE.}
}
\value{
A tibble with columns Year; Number of Negative Tests performed; Number of Positive Tests performed; Total Tests performed; Proportion of Positive Tests performed.
}
\description{
Calculate number of tests performed in an EMOD simulation
}
