% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEMODResults.R
\name{emodplot.prevalence}
\alias{emodplot.prevalence}
\title{Plot prevalence of hiv stratified by gender}
\usage{
emodplot.prevalence(data, date.start, date.end, title = "HIV prevalence")
}
\arguments{
\item{data}{A tibble returned from read.simulation.results(..., stratify_columns = c("Year", "Gender"), aggregate_columns = c("Population","Newly.Infected", "Infected")).
More aggregate columns can be used, but more stratify columns will cause problems in the plot.}

\item{date.start}{integer year to start the plotting (i.e., 2000)}

\item{date.end}{integer year to end the plotting (i.e., 2030)}
}
\value{
a ggplot with prevalence plotted
}
\description{
plots prevalence of EMOD simulations. First, it faintly plots all simulations. Then it plots a bold line of the mean of all simulations
}
