% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEMODResults.R
\name{plot.by_gender}
\alias{plot.by_gender}
\title{Plot by gender}
\usage{
\method{plot}{by_gender}(data, date.start, date.end, col2plot, title = "", unit = "")
}
\arguments{
\item{data}{A tibble returned from read.simulation.results(..., stratify_columns = c("Year", "Gender"), aggregate_columns = c("Population","Newly.Infected", "Infected")).
More aggregate columns can be used, but more stratify columns will cause problems in the plot.}

\item{date.start}{integer year to start the plotting (i.e., 2000)}

\item{date.end}{integer year to end the plotting (i.e., 2030)}

\item{col2plot}{string of column name you want to plot (i.e., "Infected")}

\item{title}{string of plot title (i.e., "Infected Plot")}

\item{unit}{string of the units for col2plot (i.e., "(Number of People)")}
}
\value{
a ggplot with all data plotted
}
\description{
Plot by gender
}
\details{
plots a metric in data. First, it faintly plots all simulations. Then it plots a bold line of the mean of all simulations
}
