% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadEMODResults.R
\name{read.simulation.results}
\alias{read.simulation.results}
\title{Read results of an EMOD simulation}
\usage{
read.simulation.results(
  results_path,
  scenario_name,
  summarize_columns = c("Newly.Infected", "Newly.Tested.Positive",
    "Newly.Tested.Negative", "Population", "Infected", "On_ART", "Died", "Died_from_HIV",
    "Tested.Past.Year.or.On_ART", "Tested.Ever", "Diagnosed"),
  stratify_columns = c("Year", "Gender"),
  min_age_inclusive = 15,
  max_age_inclusive = 49
)
}
\arguments{
\item{results_path}{string pointing to the folder which contains the ReportHIVByAgeAndGender.csv files}

\item{scenario_name}{string for the name of the scenario being read. For example, you might use "baseline" for the baseline scenario.}

\item{summarize_columns}{a vector of strings containing names of columns to be aggregated via summation. Note that spaces in column names are replaced by a period ("."). For example, "Newly Infected" becomes "Newly.Infected".}

\item{stratify_columns}{a vector of strings containing names of columns by which we will stratify the data. For example, we might want to have a separate row in the dataset for each year, so we would set stratify_columns = c("Year")}

\item{min_age_inclusive}{an integer representing the minimum age to keep while reading the data (all ages below will be filtered out)}

\item{max_age_inclusive}{an integer representing the maximum age to keep while reading the data (all ages above will be filtered out)}
}
\value{
A tibble with columns incidence and Year
}
\description{
Read results of an EMOD simulation
}
\details{
The results of an EMOD simulation are stored in a series of csv files titled "ReportHIVByAgeAndGender.csv". One of these files exists for each
simulation run (typically 250 files). This function reads and aggregates those files into a single tibble
}
