% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportEMODResults.R
\name{report.calibration.results}
\alias{report.calibration.results}
\title{Report calibration fit of EMOD simulation}
\usage{
report.calibration.results(
  experiment_path,
  ingest_file_path,
  pop_scaling_factor = 1,
  figure_path = "./",
  completed_run = TRUE
)
}
\arguments{
\item{experiment_path}{string path on big purple where experiment was run. Typically this will look like}

\item{ingest_file_path}{string path to ingest file}

\item{pop_scaling_factor}{float population scaling factor. If < 1, it will be inverted.}

\item{figure_path}{string folder where the figures generated by this function will be saved}

\item{completed_run}{bool if TRUE (default), will expect all ReportHIVByAgeAndGender.csv files to be in
the same folder. If FALSE, expects a big purple simulation folder}
}
\description{
Report calibration fit of EMOD simulation
}
