% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportEMODResults.R
\name{report.calibration.results}
\alias{report.calibration.results}
\title{Report calibration fit of EMOD simulation}
\usage{
report.calibration.results(
  experiment_path,
  ingest_path,
  pop_scaling_factor = 1,
  figure_path = "./"
)
}
\arguments{
\item{experiment_path}{path on big purple where experiment was run. Typically this will look like
/gpfs/scratch/kaftad01/experiments/Zambia--0.02--rep7--testCalibrateSeedYear_iter32___2022_03_21_12_39_07_01727}
}
\description{
Report calibration fit of EMOD simulation
}
